combine map controls
